+++ build_bare_verify/CTestTestfile.cmake
@@ -1,9 +1,0 @@
-# CMake generated Testfile for 
-# Source directory: /home/andrii/Projects/iolinki
-# Build directory: /home/andrii/Projects/iolinki/build_bare_verify
-# 
-# This file includes the relevant testing commands required for 
-# testing this directory and lists subdirectories to be tested as well.
-subdirs("examples/simple_device")
-subdirs("examples/bare_metal_app")
-subdirs("tests")