+++ build_linux/CMakeFiles/iolinki.dir/src/platform.c.o.d
@@ -1,3 +1,0 @@
-CMakeFiles/iolinki.dir/src/platform.c.o: \
- /home/andrii/Projects/iolinki/src/platform.c /usr/include/stdc-predef.h \
- /home/andrii/Projects/iolinki/include/iolinki/platform.h