+++ build_zephyr_verify/mock_kernel_entry.c
@@ -1,1 +1,0 @@
-void _mock_entry(void){}